.gitignore file was updated. HeroesScreen.kt was renamed to AppNavGraph.kt. Data package was renamed to model. In nestedScrollConnection (ChooseHeroScreen fun) override fun onPreScroll was changed to onPastScroll
