+++ .gitignore
@@ -8,13 +8,9 @@ gen/
 Thumbs.db
 *.iml
 .gradle
+gradle
 /local.properties
-/.idea/caches
-/.idea/libraries
-/.idea/modules.xml
-/.idea/workspace.xml
-/.idea/navEditor.xml
-/.idea/assetWizardSettings.xml
+/.idea
 .DS_Store
 /build
 /captures
