+++ app/src/main/java/com/example/marvel_app_project/MainActivity.kt
@@ -26,7 +26,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    HeroesApp()
+                    AppNavGraph()
                 }
             }
         }
