+++ app/src/main/java/com/example/marvel_app_project/assets/HeroesData.kt
@@ -1,7 +1,7 @@
 package com.example.marvel_app_project.assets
 
 import com.example.marvel_app_project.R
-import com.example.marvel_app_project.data.Heroes
+import com.example.marvel_app_project.models.Heroes
 
 object SampleData{
     val heroesSample = listOf(
