+++ app/src/main/java/com/example/marvel_app_project/ui/pages/ChooseHero.kt
@@ -27,7 +27,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.marvel_app_project.R
 import com.example.marvel_app_project.assets.SampleData
-import com.example.marvel_app_project.data.Heroes
+import com.example.marvel_app_project.models.Heroes
 import com.example.marvel_app_project.ui.components.ChooseHeroHeader
 import com.example.marvel_app_project.ui.components.HeroCard
 import com.example.marvel_app_project.ui.theme.Marvel_app_projectTheme
@@ -46,7 +46,8 @@ fun ChooseHeroScreen(onHeroImageTaped:(Heroes) -> Unit) {
     }
     val nestedScrollConnection = remember {
         object : NestedScrollConnection{
-            override fun onPreScroll(
+            override fun onPostScroll(
+                consumed: Offset,
                 available: Offset,
                 source: NestedScrollSource
             ): Offset {
