+++ app/src/main/java/com/example/marvel_app_project/ui/pages/SingleHero.kt
@@ -18,7 +18,7 @@ import androidx.compose.ui.res.stringResource
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.example.marvel_app_project.R
-import com.example.marvel_app_project.data.Heroes
+import com.example.marvel_app_project.models.Heroes
 import com.example.marvel_app_project.ui.components.SingleHeroTextField
 import com.example.marvel_app_project.ui.theme.Sizes
 import com.example.marvel_app_project.ui.theme.Spaces
